add ssh-agent auth
Instead of prompting for a password when executing runCmd, we
will instead use ssh-agent auth. If the SSH_AUTH_SOCK env
variable is empty, then we will use a PasswordCallback to prompt
for the password.

If we want to send the password via stdin to execute a command
with sudo we will still prompt for the password.